Move forms and widget
- Move forms to separate files
- Move custom widget to separate file
