+++ geekhunter/company_app/forms/AdminCompanyCardForm.py
@@ -1,0 +1,16 @@
+from django import forms
+
+from ..models import Card
+from .widgets import BigTextarea
+
+
+class AdminCompanyCardFrom(forms.ModelForm):
+    class Meta:
+        model = Card
+        widgets = {
+            'about': BigTextarea,
+            'priorities': BigTextarea,
+            'awards': BigTextarea,
+            'error_text': BigTextarea
+        }
+        fields = '__all__'
