+++ geekhunter/company_app/forms/AdminVacancyForm.py
@@ -1,0 +1,14 @@
+from django import forms
+
+from ..models import Vacancy
+from .widgets import BigTextarea
+
+
+class AdminVacancyFrom(forms.ModelForm):
+    class Meta:
+        model = Vacancy
+        widgets = {
+            'description': BigTextarea,
+            'error_text': BigTextarea
+        }
+        fields = '__all__'
